+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - Domains allow variables (dropdown, radio, scale, etc) to share common choices
 - Experimental reporter now formats calculated variables based on the variables' format
 - Uploaded file exports are now always zipped even if only one file has been uploaded
+- Variable domains can now only be changed if the newly selected domain options contain values for each previously captured response
 
 ### Bug Fix
 - Hitting `p` no longer triggers switching to the global search when focused on a link or a drop down menu